+++ CHANGES.rst
@@ -6,7 +6,7 @@ Changelog
 
 New:
 
-- extended step to501 with reindexing get_icon (#1226)
+- extended step to501 with reindexing getIcon (#1226)
   [fgrcon]
 
 - Removed fake kupu tool and related settings and resources.